Move selectors into container subfolder
