+++ ui/apps/platform/cypress/integration/accessControl/accessControlRoles.test.js
@@ -1,5 +1,3 @@
-import { selectors } from '../../constants/AccessControlPage';
-
 import withAuth from '../../helpers/basicAuth';
 
 import {
@@ -11,6 +9,7 @@ import {
     visitAccessControlEntitiesWithStaticResponseForPermissions,
     visitAccessControlEntity,
 } from './accessControl.helpers';
+import { selectors } from './accessControl.selectors';
 
 const defaultNames = ['Admin', 'Analyst', 'Continuous Integration', 'None', 'Sensor Creator'];
